requirements.txt
arabic-reshaper
pandas
tabulate
bidi.algorithm